infragraph/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "infragraph"
version = "0.1.0"
description = "Map your homelab infrastructure from config files to a readable graph."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "The Nerd Hole" }]
keywords = ["infrastructure", "homelab", "network", "graph", "docker", "nginx"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-snapshot>=0.9",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
infragraph = "infragraph.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
